修复无法存档与读取的问题

+++ Assets/Scripts/Chat/ParagraphData.cs
@@ -7,7 +7,7 @@ public class ParagraphData
     public float ParagraphPriority;//段落优先级
     public string ParagraphCondition;//触发段落方法
 
-    public ParagraphData(string paragraphName, int paragraphPriority,string paragraphCondition)
+    public ParagraphData(string paragraphName, float paragraphPriority,string paragraphCondition)
     {
         ParagraphName=paragraphName;
         ParagraphPriority=paragraphPriority;
